+++ src/Mvc/AscmvcEventManager.php
@@ -5,7 +5,7 @@
  * @package    LightMVC/ASCMVC
  * @author     Andrew Caya
  * @link       https://github.com/lightmvc/ascmvc
- * @version    2.0.0
+ * @version    2.0.1
  * @license    http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0.
  * @since      2.0.0
  */
@@ -15,6 +15,8 @@ namespace Ascmvc\Mvc;
 use Zend\Diactoros\Response;
 use Zend\EventManager\EventManager;
 use Zend\EventManager\SharedEventManagerInterface;
+
+use function Ascmvc\getNamespaceFromPath;
 
 /**
  * Class AscmvcEventManager
@@ -87,11 +89,9 @@ class AscmvcEventManager extends EventManager
                 $filePath = $fileInfo->getPathName();
                 require_once $filePath;
 
-                $filePathArray = explode('/', $filePath);
-                $fileName = array_pop($filePathArray);
-                array_pop($filePathArray);
-                $domainName = array_pop($filePathArray);
-                $controllerName = $domainName. '\\Controllers\\' . substr($fileName, 0, strlen($fileName) - 4);
+                $filePathArray = getNamespaceFromPath($filePath);
+
+                $controllerName = $filePathArray['domainName'] . '\\Controllers\\' . substr($filePathArray['fileName'], 0, strlen($filePathArray['fileName']) - 4);
 
                 $response = $controllerName::onBootstrap($event);